change Matrix datatype to float & add matrixC parameter

+++ MatrixMultiplikation_java/src/matrixmult/Matrix.java
@@ -3,11 +3,11 @@ package matrixmult;
 import java.util.Random;
 
 public class Matrix {
-    private int[][] matrix;
+    private float[][] matrix;
 
     Matrix(int n, int m) throws Exception {
         if(n == 0 || m == 0) throw new Exception("n = 0 or m = 0 is not permitted");
-        matrix = new int[n][m];
+        matrix = new float[n][m];
     }
 
     public void randomize() {
@@ -20,7 +20,7 @@ public class Matrix {
     }
 
     public void print() {
-        for (int[] rows : matrix) {
+        for (float[] rows : matrix) {
             for (int c = 0; c < matrix[0].length; c++)
                 System.out.print(rows[c] + " ");
             System.out.println();
@@ -28,7 +28,7 @@ public class Matrix {
         System.out.println();
     }
 
-    public int[][] getMatrix() {
+    public float[][] getMatrix() {
         return matrix;
     }
 }
